+++ selectel/resource_selectel_mks_nodegroup_v1.go
@@ -11,6 +11,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	"github.com/selectel/go-selvpcclient/selvpcclient/resell/v2/quotas"
 	"github.com/selectel/go-selvpcclient/selvpcclient/resell/v2/tokens"
 	v1 "github.com/selectel/mks-go/pkg/v1"
 	"github.com/selectel/mks-go/pkg/v1/nodegroup"
@@ -112,7 +113,7 @@ func resourceMKSNodegroupV1() *schema.Resource {
 			"taints": {
 				Type:     schema.TypeList,
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"key": {
@@ -224,6 +225,15 @@ func resourceMKSNodegroupV1Create(ctx context.Context, d *schema.ResourceData, m
 		KeypairName:      d.Get("keypair_name").(string),
 		AffinityPolicy:   d.Get("affinity_policy").(string),
 		AvailabilityZone: d.Get("availability_zone").(string),
+	}
+
+	projectQuotas, _, err := quotas.GetProjectQuotas(ctx, resellV2Client, d.Get("project_id").(string))
+	if err != nil {
+		return diag.FromErr(errGettingObject(objectProjectQuotas, d.Get("project_id").(string), err))
+	}
+
+	if err := checkQuotasForNodegroup(projectQuotas, createOpts); err != nil {
+		return diag.FromErr(errCreatingObject(objectNodegroup, err))
 	}
 
 	// Check nodegroup autoscaling options.
@@ -386,6 +396,12 @@ func resourceMKSNodegroupV1Update(ctx context.Context, d *schema.ResourceData, m
 		hasChanged = true
 	}
 
+	if d.HasChange("taints") {
+		taints := d.Get("taints").([]interface{})
+		updateOpts.Taints = expandMKSNodegroupV1Taints(taints)
+		hasChanged = true
+	}
+
 	if d.HasChanges("enable_autoscale", "autoscale_min_nodes", "autoscale_max_nodes") {
 		enableAutoscale := d.Get("enable_autoscale").(bool)
 		autoscaleMinNodes := d.Get("autoscale_min_nodes").(int)
@@ -412,6 +428,28 @@ func resourceMKSNodegroupV1Update(ctx context.Context, d *schema.ResourceData, m
 	}
 
 	if d.HasChange("nodes_count") {
+		oldValue, newValue := d.GetChange("nodes_count")
+		newNodesCount := newValue.(int) - oldValue.(int)
+
+		newNodesRequest := nodegroup.CreateOpts{
+			Count:            newNodesCount,
+			CPUs:             d.Get("cpus").(int),
+			RAMMB:            d.Get("ram_mb").(int),
+			VolumeGB:         d.Get("volume_gb").(int),
+			VolumeType:       d.Get("volume_type").(string),
+			LocalVolume:      d.Get("local_volume").(bool),
+			AvailabilityZone: d.Get("availability_zone").(string),
+		}
+
+		projectQuotas, _, err := quotas.GetProjectQuotas(ctx, resellV2Client, d.Get("project_id").(string))
+		if err != nil {
+			return diag.FromErr(errGettingObject(objectProjectQuotas, d.Get("project_id").(string), err))
+		}
+
+		if err := checkQuotasForNodegroup(projectQuotas, &newNodesRequest); err != nil {
+			return diag.FromErr(errUpdatingObject(objectNodegroup, d.Id(), err))
+		}
+
 		resizeOpts := nodegroup.ResizeOpts{
 			Desired: d.Get("nodes_count").(int),
 		}